Changing reliabilityFactor to Double instead Int

+++ mvvm-coroutines-clean-architecture/app/src/main/java/br/com/mobiplus/gitclient/presentation/ui/main/model/GitRepoUIModel.kt
@@ -8,7 +8,7 @@ data class GitRepoUIModel(
     var description: String? = null,
     var stargazersCount: Int? = null,
     var forksCount: Int? = null,
-    var reliabilityFactor: Int? = null,
+    var reliabilityFactor: Double? = null,
     var language: String? = null,
     var openIssuesCount: Int? = null,
     var ownerLogin: String? = null,
